+++ src/d2_docker/utils.py
@@ -10,13 +10,12 @@ import tempfile
 import time
 import yaml
 import urllib.request
-import json
-import zipfile
 from setuptools._distutils import dir_util
 from pathlib import Path
 from typing import Optional
 
 import d2_docker
+from d2_docker.glowroot import get_glowroot_zip, get_port_glowroot
 from .image_name import ImageName
 
 PROJECT_NAME_PREFIX = "d2-docker"
@@ -40,16 +39,6 @@ def get_dhis2_war_url(version):
         else "{}/dhis2-stable-{}.war".format(short_version, version)
     )
     return releases_base_url + "/" + path
-
-def get_latest_glowroot_url():
-    glowroot_releases_url = "https://api.github.com/repos/glowroot/glowroot/releases/latest"
-    glowroot_download_url = "https://github.com/glowroot/glowroot/releases/download"
-    with urllib.request.urlopen(glowroot_releases_url) as response:
-        data = response.read().decode()
-        release_info = json.loads(data)
-
-    tag_name = release_info["tag_name"]
-    return "{}/{}/glowroot-{}-dist.zip".format(glowroot_download_url, tag_name, tag_name.lstrip("v"))
 
 
 def docker_build(directory, tag):
@@ -291,24 +280,6 @@ def run_docker_compose(
     core_image_name = core_image or get_core_image_name(data_image)
     post_sql_dir_abs = get_absdir_for_docker_volume(post_sql_dir)
     scripts_dir_abs = get_absdir_for_docker_volume(scripts_dir)
-    glowroot_path=None
-
-    if args[0] == "up":
-        glowroot_file = tempfile.NamedTemporaryFile(delete=False, prefix="glowroot_", suffix=".zip", dir="/tmp")
-        glowroot_path = glowroot_file.name
-
-        atexit.register(lambda: os.remove(glowroot_path) if os.path.exists(glowroot_path) else None)
-        if glowroot_zip:
-            logger.debug("Copy zip file: {} -> {}".format(glowroot_zip, glowroot_path))
-            shutil.copy(glowroot_zip, glowroot_path)
-        elif glowroot:
-            glowroot_url = get_latest_glowroot_url()
-            logger.info("Download file: {}".format(glowroot_url))
-            urllib.request.urlretrieve(glowroot_url, glowroot_path)
-        else:
-            # empty zipfile
-            with zipfile.ZipFile(glowroot_path, mode="w") as zf:
-                pass
 
     env_pairs = [
         ("DHIS2_DATA_IMAGE", final_image_name),
@@ -330,8 +301,8 @@ def run_docker_compose(
         # Add ROOT_PATH from environment (required when run inside a docker)
         ("ROOT_PATH", ROOT_PATH),
         ("PSQL_ENABLE_QUERY_LOGS", "") if not enable_postgres_queries_logging else None,
-        ("GLOWROOT_PORT", "{}:4000".format(glowroot_port) if glowroot_port else "4000:4000") if (glowroot or glowroot_zip) else ("GLOWROOT_PORT", None),
-        ("GLOWROOT_ZIP", get_absfile_for_docker_volume(glowroot_path)),
+        ("GLOWROOT_PORT", get_port_glowroot(glowroot_port, glowroot_zip, glowroot)),
+        ("GLOWROOT_ZIP", get_glowroot_zip(args[0], glowroot_zip, glowroot)),
     ]
     env = dict((k, v) for (k, v) in [pair for pair in env_pairs if pair] if v is not None)
